backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openpyxl==3.11.0
pandas==2.1.3
numpy==1.26.2
boto3==1.29.7
supabase==2.3.5
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
uuid==1.30
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2
